# Allow passing Baseline and Candidate paths via script args