--- a/src/python/competition.py
+++ b/src/python/competition.py
@@ -216,7 +216,7 @@ class Index(object):
       return self.assignedName
 
     name = [self.dataSource.name,
-            self.checkout]
+            benchUtil.checkoutToName(self.checkout)]
 
     if self.extraNamePart is not None:
       name.append(self.extraNamePart)
@@ -469,6 +469,7 @@ class Competition(object):
     for c in self.competitors:
       if c.name == name:
         raise RuntimeError(f'competitor named {name} already added')
+    print('Using checkout:[%s] for competitor:[%s]' % (benchUtil.checkoutToPath(checkout), name))
     c = Competitor(name, checkout, **kwArgs)
     c.competition = self
     self.competitors.append(c)