--- a/src/python/example.py
+++ b/src/python/example.py
@@ -25,14 +25,19 @@ if __name__ == '__main__':
                                    description='Run a local benchmark on provided source dataset.')
   parser.add_argument('-s', '-source', '--source',
                       help='Data source to run the benchmark on.')
-  parser.add_argument('-c', '-concurrentSearches', '--concurrentSearches', action='store_true',
+  parser.add_argument('-concurrentSearches', '--concurrentSearches', action='store_true',
                       help='Run concurrent searches')
+  parser.add_argument('-b', '--baseline', default='lucene_baseline',
+                      help='Path to lucene repo to be used for baseline')
+  parser.add_argument('-c', '--candidate', default='lucene_candidate',
+                      help='Path to lucene repo to be used for candidate')
   args = parser.parse_args()
+  print('Running benchmarks with the following args: %s' % args)
 
   sourceData = competition.sourceData(args.source)
   comp =  competition.Competition()
 
-  index = comp.newIndex('lucene_baseline', sourceData,
+  index = comp.newIndex(args.baseline, sourceData,
                         addDVFields = True,
                         facets = (('taxonomy:Date', 'Date'),
                                   ('taxonomy:Month', 'Month'),
@@ -44,14 +49,14 @@ if __name__ == '__main__':
                                   ('sortedset:RandomLabel', 'RandomLabel')))
 
   # create a competitor named baseline with sources in the ../trunk folder
-  comp.competitor('baseline', 'lucene_baseline',
+  comp.competitor('baseline', args.baseline,
                   index = index, concurrentSearches = args.concurrentSearches)
 
   # use the same index here
-  # create a competitor named my_modified_version with sources in the ../patch folder
+  # create a competitor named my_modified_version (or provided candidate name) with sources in the ../patch folder
   # note that we haven't specified an index here, luceneutil will automatically use the index from the base competitor for searching 
   # while the codec that is used for running this competitor is taken from this competitor.
-  comp.competitor('my_modified_version', 'lucene_candidate',
+  comp.competitor('my_modified_version', args.candidate,
                   index = index, concurrentSearches = args.concurrentSearches)
 
   # start the benchmark - this can take long depending on your index and machines